Server/requirements.txt
asyncio
websockets
opencv-python
pillow
numpy
ultralytics
paddleocr